+++ src/Testing/ZopeTestCase/threadutils.py
@@ -10,56 +10,15 @@
 # FOR A PARTICULAR PURPOSE.
 #
 ##############################################################################
-"""Parts of ZServer support are in this module so they can
-be imported more selectively.
-"""
 
-from threading import Thread
-from StringIO import StringIO
+from zope.deferredimport import deprecated
 
-dummyLOG = StringIO()
-
-
-def setNumberOfThreads(number_of_threads):
-    '''Sets number of ZServer threads.'''
-    try:
-        from ZServer.Zope2.Startup.config import setNumberOfThreads
-        setNumberOfThreads(number_of_threads)
-    except ImportError:
-        pass
-
-
-def zserverRunner(host, port, log=None):
-    '''Runs an HTTP ZServer on host:port.'''
-    from ZServer import logger, asyncore
-    from ZServer import zhttp_server, zhttp_handler
-    if log is None: log = dummyLOG
-    lg = logger.file_logger(log)
-    hs = zhttp_server(ip=host, port=port, resolver=None, logger_object=lg)
-    zh = zhttp_handler(module='Zope2', uri_base='')
-    hs.install_handler(zh)
-    asyncore.loop()
-
-
-class QuietThread(Thread):
-    '''This thread eats all exceptions'''
-    def __init__(self, target=None, args=(), kwargs={}):
-        Thread.__init__(self, target=target, args=args, kwargs=kwargs)
-        self.__old_bootstrap = Thread._Thread__bootstrap
-    def __bootstrap(self):
-        try: self.__old_bootstrap(self)
-        except: pass
-    _Thread__bootstrap = __bootstrap
-
-
-def QuietPublisher(self, accept):
-    '''This server eats all exceptions'''
-    try: self.__old_init__(accept)
-    except: pass
-
-
-from ZServer.PubCore.ZServerPublisher import ZServerPublisher
-if not hasattr(ZServerPublisher, '__old_init__'):
-    ZServerPublisher.__old_init__ = ZServerPublisher.__init__
-    ZServerPublisher.__init__ = QuietPublisher
-
+# BBB Zope 5.0
+deprecated(
+    'Please import from ZServer.Testing.threadutils.',
+    dummyLOG='ZServer.Testing.threadutils:dummyLOG',
+    setNumberOfThreads='ZServer.Testing.threadutils:setNumberOfThreads',
+    zserverRunner='ZServer.Testing.threadutils:zserverRunner',
+    QuietThread='ZServer.Testing.threadutils:QuietThread',
+    QuietPublisher='ZServer.Testing.threadutils:QuietPublisher',
+)